OCSADV-272 Reduce the amount of methods in GemsUtils4Java moving them as necessary #time 30m
