+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/tpe/GemsGuideStarWorker.java
@@ -195,11 +195,11 @@ public class GemsGuideStarWorker extends SwingWorker implements MascotProgress {
             return Double.compare(a1.toDegrees(), a2.toDegrees());
         });
 
-        posAngles.add(GemsUtils4Java.toNewAngle(obsContext.getPositionAngle()));
-        posAngles.add(GemsUtils4Java.toNewAngle(new Angle(0., Angle.Unit.DEGREES)));
-        posAngles.add(GemsUtils4Java.toNewAngle(new Angle(90., Angle.Unit.DEGREES)));
-        posAngles.add(GemsUtils4Java.toNewAngle(new Angle(180., Angle.Unit.DEGREES)));
-        posAngles.add(GemsUtils4Java.toNewAngle(new Angle(270., Angle.Unit.DEGREES)));
+        posAngles.add(ModelConverters.toNewAngle(obsContext.getPositionAngle()));
+        posAngles.add(ModelConverters.toNewAngle(new Angle(0., Angle.Unit.DEGREES)));
+        posAngles.add(ModelConverters.toNewAngle(new Angle(90., Angle.Unit.DEGREES)));
+        posAngles.add(ModelConverters.toNewAngle(new Angle(180., Angle.Unit.DEGREES)));
+        posAngles.add(ModelConverters.toNewAngle(new Angle(270., Angle.Unit.DEGREES)));
         return posAngles;
     }
